Update go_to_goal.py
Update Version 3: Solve the problem with the first spawned turtle. 
Previously, when the first goal turtle is spawned, the moving turtle moves towards a wrong goal position (where there's no turtle at all) instead of moving towards the goal turtle. And when it reached its goal, the goal turtle is killed although it have not be reached.
I believe this is because when turtlesim is run, a turtle (turtle2) is created and displayed. Then our program creates another turtle (turtle3, 4, 5...) with spawn, and it is set as the goal, yet not displayed.
My solution is killing the turtle generated by turtlesim (turtle2), and then spawn a turtle (turtle3) so that the new turtle is set as goal and displayed.